Add dependency on various libraries from ubuntu-restricted-extras.
This is a subset of the dependencies from ubuntu-restricted-extras, including its dependent metapackage ubuntu-restricted-addons (which is installed by default with Ubuntu when the installer checks the box for the third party codec addons).  We include everything except for the MS fonts (which requires the user to "sign" and end user license agreement) and flash (which currently is causing problems for us in Epiphany).  Among other potential improvements, this fixes known problems playing back certain MP3 files.

[endlessm/eos-shell#486]