eos-convert-system: Transition to a more complete Debian system
When doing eos-dev-fix the system is essentially put in a full
read-write mode, which violates the assumptions of ostree (the system
being read-only). Violating this assumption means the ostree object
store cannot be trusted anymore (an ostree system is hardlink farm
directly into the object store, thus doing changes into the system
directly changes the object store).

Given all this, it would be better to do a more complete transition to
a more general Debian system such that e.g. kernel upgrade work and
avoid the r/o mounted /usr on reboot (which breaks dpkg when trying it
after boot).

This does just that. Note that it does keep the /ostree infrastructure
in place and does a hardlink copy instead of a regular one to not double
the disk usage (which might be awkward in some cases).