eos-dev-fix: Disable upgrades and bail if one was pending
Disable automatic system updates before doing any other modification to
the running system, once the modifications are done there is no way to
still safely update the system.

Also, prevent any real modifications to be done if a system upgrade was
already prepared and put into place for the next boot. As in that case
doing modifications onto the current system will result in a very
confused system after reboot