Add more Xorg drivers for x86
With ati, intel and nouveau, we cover the vast majority of hardware
cases. vmware allows more resolutions and better performance in a VMWare
vm. Finally, modesetting is like vesa for the KMS world. Any time a new
KMS device shows up in the kernel, modesetting provides a generic
interface to use it from X. In the worst case, X will fail to start if a
KMS driver is loaded and there's no KMS capable X driver.

[endlessm/eos-shell#2220]