Install the radeon Xorg driver, not the ati metapackage
xserver-xorg-video-ati is actually a metapackage that also requires the
mach64 and r128 drivers. Not only do we not have those drivers, but
they're only needed for ancient hardware. The driver we really want that
covers any ATI hardware from the past 10ish years is the radeon driver
included in the xserver-xorg-video-ati source package.

[endlessm/eos-shell#2220]