Add libhal1-flash for x86 flash DRM content
In order to play DRM flash content such as Amazon videos, the flash
DRM capabilities need to be enabled. The hal-flash library does this
by providing the needed symbols from libhal1 and using udisks2 to
query for information.

This is only enabled on i386 as we have currently have no way to
enable flash on ARM.

[endlessm/eos-shell#556]