Break unwanted links of empty files after copying from ostree
OSTree has a clever optimization that files of the same size can be
hardlinked together to save space. This is OK when everything's read
only, but after running eos-convert-system any of those files that gets
written to affects all of them.

The case that we really want to check for is empty files. This is most
likely to be where OSTree has performed this optimization. Other cases
where files are of the same size and hardlinked are likely to be
intentional, such as the hardlinks between bzip2, bunzip2 and bzcat.

Add a step in convert-system after copying the files to /sysroot that
finds empty files with multiple links. Then pass that list of files to a
new script, eos-break-links, that will perform the dance of replacing
the file with an identical one with a new inode.

eos-break-links also has a --zero mode that allows someone with a
corrupted system to find the hardlinked files, replace them as above,
and truncate the new file.

[endlessm/eos-shell#2251]