Narrow the locations to break links in
I forgot that /sysroot contains the real ostree directory and other
things that were not part of the deployment copy. Letting
eos-break-links run through the ostree object database probably corrupts
it. While it's probably not an issue to do this and also break links in
other locations, it would be best to only operate in the copied
directories.