Add games to apps list
Added 14 of the "top 15" games identified by Patrick
to the apps dependencies.  These are not installed by default
in the image, but this may be convenient for QA testing
until the app manager is in place.

Excluded extreme tux racer from arm, as it is known to not work
properly on the Odroid.  The others all built successfully but
have not yet been tested on the Odroid.

[endlessm/eos-shell#2465]