Temporarily remove disk health check
As currently implemented, this relies on smartmontools,
which brings in smartd and postfix, which we don't really
need or want in our default image.

In the future, we will either separate out the smartctl
binary that we actually want, or find a way to do this
without bringing in any further dependencies
(such as how gnome-disks accesses the SMART disk data).

[endlessm/eos-shell#2841]