Add all i386 architecture specific packages to core
Currently these packages are specified in the image builder as part of
the platform definition. However, they would be needed on any Endless
i386 installation, so there's no reason not to fully specify our i386
package set in eos-core. We should always be able to use a generic i386
architecture image with no platform variants.

[endlessm/eos-shell#4107]