Add ARM platform specific core packages
Unlike i386, the ARM images are platform specific and usually require a
different package set to support that particular variant. This list has
been stored in the image builder, but it would be more robust if this
was stored as a metapackage like eos-core and not tied to the builder
itself. Furthermore, it would make determining the true set of packages
included in the base ARM images more straightforward.

This creates lists supporting two new metapackages: eos-core-odroidu2
and eos-core-sqwerty. These contain the platform specific package lists
from the image builder for each of our current ARM platforms. eos-core
has also been added to the list to support a simple "apt-get install
eos-core-$platform" providing all needed packages on ARM.

[endlessm/eos-shell#4107]