Add eos-pepflashplugin-updater to core
Now that we have upgraded chromium-browser from v33 to v40,
we need to use Pepper Flash from Google rather than the Adobe version.
For now, we can keep both versions of the Flash plugin updater
installed, as the Adobe version may be useful for other situations
(e.g., if we make Firefox available).

[endlessm/eos-shell#4146]