eos-convert-system-qa: Use ostree admin set-origin
Stable version images are deployed with an ostree branch of
eos$major/$arch so that we can automatically upgrade users across minor
versions. When converting to the staging/dev server, not only does the
URL need to change, but the branch to. This means that not only the repo
config file, but also the deployment's origin file needs to be updated.

Use the new ostree admin set-origin builtin to manage this change. It
knows how to update the origin file for the current deployment as well
as the url and branches in the repo config file. We do need some
additional calculations to determine the correct branch, though.

[endlessm/eos-shell#4845]