tech-support: Exit on errors for most scripts
Most of the tech support scripts do dangerous operations and should not
continue if there's an error. Only eos-launch-all is currently handling
all errors. All other scripts should use bash's errexit feature to exit
the program if any of the commands return an error.

In order to handle that correctly, some fixes are needed. Errors need to
be caught rather than checking $?, and cleanup actions need to be
collected in an EXIT trap.

[endlessm/eos-shell#5189]