Update eos-usb-image for GPT partitioning
We no longer have a separate /boot partition,
and we no longer use a special partition 4 as a marker.

Also, swap is no longer managed via fstab,
and it is disabled by default if less than 100 GB available,
so no special handling is necessary to disable swap
on USB thumb drives up to 64 GB.

For Intel, the root partition number has moved from 3 to 2,
but it is 1 on ARM.  We replace the hard-coded partition number
with a query for the device with the label "ostree".

[endlessm/eos-shell#5413]