eos-diagnostics: fix application list
The app manager DBus method no longer exists.  There is a DBus method
to the app store, but it does not provide all the information that
we want.  Instead, let's use the eamctl command for the app manager,
which has a "--verbose" option that includes the information we want
(app ID and version).

[endlessm/eos-shell#5797]