Install different defaults.list for ARM and i386 products
This removes defaults-list from the install list for eos-core and create
separate lists for eos-core-i386 and eos-core-ec100, which copy
different defaults.list for each package.

On ARM we want to have H.264 and webm videos opened in the browser by
default, since the performance is much better than in totem at this
moment. We may revert this if we can improve the gstreamer performance
for these kinds of video.

[endlessm/eos-shell#5892]