Do not install Empathy or Telepathy
Remove Empathy as per #5683. Remove Telepathy as well -- it will get
pulled in if the user installs Empathy manually. Also, remove the
Telepathy Folks plugin. It seemed undesirable to keep Telepathy around
by default just for this plugin. I can't think of why it would be
useful, except for GNOME Contacts, which is not installed.

[endlessm/eos-shell#5683]