Remove eos-reset-password script
This was a bit of a hack, in that it assumes a single-disk image
on /dev/sda, and did not reset the user's keyring.  We now have
a fully supported password reset feature from the login screen.

Also, remove the related eos-list-users script, which was also
a bit of a hack and no longer needed with the new feature.

[endlessm/eos-shell#6324]