Revert "Install different defaults.list for ARM and i386 products"
We have improved the performance issue in totem which had initially led
to the split of mime types for armhf vs. i386.  While the video
performance is still slightly better in chromium than in totem,
the overall user experience is better in totem, so we will revert
to opening videos in totem by default.

Since there are no longer any differences in the handling on armhf
vs. i386, let's manage just a single copy of the defaults list.

This reverts commit e62b8c2182d8998819020fa00faf5bcbd83561cf.

[endlessm/eos-shell#6057]