Add dependency on libglu1-mesa to xdg-app runtime
It is required by several packages and was already part of eos-core,
so it makes sense this becomes part of the runtime.

In the mid-term, this should not be here but in a runtime extension,
so it would be possible to load different GL libs depending on the
hardware platform.