Add eos-installer-meta metapackage
This will be used for a minimal USB OS/image for flashing Endless images
to disks. The idea is to have a minimal desktop that starts directly
into a flashing tool. This is a cut down eos-core with almost all of the
explicit library dependencies have been removed as it's expected that
they'll be pulled in as necessary by other packages.

The eos-installer-meta name is used to prevent a collision with the
likely eos-installer package that will provide the actual flashing tool.

This is a backport of e086088 without the refactoring to split up the
depends files into common subsets.

https://phabricator.endlessm.com/T11283