Factor out core and runtime common dependencies
Add 2 include files, app-depends and base-depends to be shared between
the core and runtime metapackages. This will prevent having to add
duplicate dependencies betwen the core and runtime.

The app-depends file represents dependencies supporting apps rather than
the OS itself. Many of the library dependencies here are really runtime
specific and should be moved into the runtime metapackage when xdg-app
is ready.

The base-depends file represents common dependencies for any OS or
runtime.

https://phabricator.endlessm.com/T11283