platform: Drop compat libraries
These packages are only available in the shared repo because reprepro
doesn't remove them, but we don't have them for amd64. As it is, on
master the apps should be fixed up to depend on the current versions of
these libraries. Eventually the entire app-depends file will be moved
into the platform depends file when we no longer support bundles.

https://phabricator.endlessm.com/T11586