eos-convert-system: restore package layout for ARM kernels too
Now that dracut and linux-meson packages are providing binaries directly
compatible with the bootloader, we can treat ARM kernels as normal:
restore the way that they were installed from packages.

We also set up the same symlinks that linux-meson postinst does, and
make the bootloader use these links, allowing for ARM kernels to be
updated with apt.

https://phabricator.endlessm.com/T1404