Remove eos-usb-image
Starting on EOS 3.0, we are generating the root partition fs UUID in the
first boot, and dinamically passing it from the bootloader to the
kernel. This means a regular image should boot without problems from a
USB drive on systems which have EOS installed on the internal storage,
without the need to prepare it with eos-usb-image first.

https://phabricator.endlessm.com/T12004