Remove eos-resize-device
This script was made to work around an issue where the first-boot
resize & repartition process was not reliable on USB (when you might
also have another Endless instance visible on an internal HDD).

That bug is now fixed so remove the script, which consisted of some
painful duplication of initramfs code.