eos-convert-system-qa: Handle authentication for ostree dev repos
For a while, the production and demo repos have been served without
requiring an HTTP user and password. That means that when we convert to
the dev repos, this authentication needs to be added.

Prompt the user for the password since we're trying not to include it in
the OS. Since the authentication is quite specific to our ostree server,
construct the URL completely instead of trying to edit it in place.

https://phabricator.endlessm.com/T3784