eos-write-image: support writing image to stdout
This may seem a little weird -- why not just zcat the file yourself? --
but I want eos-write-live to reuse the `pv` logic here, but write the
uncompressed image to a file, not a device.

The other option would be to allow regular files as the 'device'
argument, but this way seems less error-prone: in the usual case of
writing to a device, I'd rather get an error than have a 8G file created
if I typo the device name.