eos-windows-uninstall: retrieve Windows partitions
First commit: This script will be used to restore the original bootrack after the
Winstaller is run. It takes one argument: the device which has Windows +
Endless on it.

Mount each Windows partition read only, and search for
endless/endless.img and endless/bootrack.img. Fail if we don't find
either.

Check that the partition table has not changed since the installer was
run. Check that bootrack.img exists alongside endless.img and that
boottrack.img is the same size as the space before the first partition. Write
boottrack.img to the start of the disk.