eos-windows-uninstall: sanity checks
Check that the partition tables are the same, but remove device
labels and label-ids before comparing the partition tables, so we
don't abort every time.

The device name and partition name will always differ between the
table on boottrack.img and the table on the device. Strip them out
before comparing.

First partition is not always found 1MiB from the start of the disk.
Get the start of the first partition from the partion table instead.

Since it is unsupported move the endless directory, we don't need to
use find to search for endless/endless.img. Use a simple file check
instead.

Write the first 440 bytes, not 446. Windows uses bytes 440-446 as a
signature, so avoid overwriting them.

Add option to udisksctl to mount read only.