eos-write-image: Move grubx64.efi on removable devices
For removable media we do not have fallback.efi in place. At boot
EFI/BOOT/BOOTX64.efi is loaded and being fallback.efi missing, it tries
to load grubx64.efi in the current directory as specified by
DEFAULT_LOADER at compile-time.

In the current setup for removable media we do not ship grubx64.efi
also in EFI/BOOT/, causing a boot failure. Fix this reworking the ESP
and moving everything in EFI/BOOT/ when dealing with removable devices.

Signed-off-by: Carlo Caione <carlo@endlessm.com>