eos-diagnostics: add hardware info dump mode
We currently collect hardware info manually and populate wiki pages
for each supported product.

With the increased number of 3rd party products, we need something more
scalable and we need an easy way of receiving this info from 3rd parties.

Add a new mode to eos-diagnostics which only outputs information relevant
to hardware enablement efforts, and output in a more organised format
(1 file per information item, in a tar archive).

Also include the ACPI DSDT + SSDT in the output. This requires root access.