eos-reset-password: fix for ostree systems
Somehow I must have inadvertantly tested on a non-ostree system
when I cleaned up for eos3 / split disk support, as the script
is broken when assuming that /etc is available directly under
the mount point.  Instead, bring back the logic that used ostree
admin to find the currently deployed directory.

https://phabricator.endlessm.com/T14303