Language packs: remove libreoffice-help for some languages
These packages are quite large, and for some of our less-supported
languages (for which we are less likely to have offline-only users),
let's free up some space, since these language packs are installed
in the ostree for all users.  In the future, we can revisit when
we have a way to install only those language packs that each
user needs.

We keep the offline help support for en, es, and pt-br, though,
as full offline support is important for these languages.

https://phabricator.endlessm.com/T14598