Language packs: add libreoffice l10n for de, ro, zh-tw
We are targeting support for traditional Chinese (Taiwan),
and we have active community members translating German and Romanian,
so let's include these small translation packages.

Note that we intentionally exclude the larger libreoffice help packages
for these languages until we have a better mechanism for delivering
language packs outside of the core ostree.

https://phabricator.endlessm.com/T14598