eos-download-image: Add product mirroring mode
Add a -m|--mirror option that allows mirroring an entire set of product
releases. In this mode, all versions of all personalities of a product
are downloaded, and the manifest files are written out. The full paths
of the downloaded files are used so that the tree mirrors the images
CDN.

https://phabricator.endlessm.com/T15116