eos-diagnostics: Include verbose journal output with --verbose
This includes all human-readable log fields, such as the COREDUMP_*
fields, which associate journal entries with coredump files; and the
CODE_* fields, which point to specific lines and functions relevant to a
journal entry.

See `man systemd.journal-fields`.

This is normally outputted if a new --verbose argument is passed to
eos-diagnostics; or if outputting coredump data. Otherwise, journal
output is outputted as before.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15177