eos-diagnostics: Include coredumps in hardware tarball
Include a list of coredumps in the log file output, and the core files
themselves in the tarball. This is only an option when running with the
--hardware option, as it requires root privileges.

If core files are stored externally (/var/lib/systemd/coredump/) they
will be copied into the tarball. This is the default in coredump.conf
(see `man 5 coredump.conf`). If they are included in the journal
(`Storage=journal` in coredump.conf), they will not be included in the
output, since we only log human-readable journal keys. However, that’s
not how we have systemd-coredump configured, so it’s not a problem.

The files are prefixed with `coredump_`.

This will be useful on systems (such as converted ones) where coredumps
are enabled; not so useful on systems where they are not enabled. But
not harmful.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15177