write-live-image: add Joliet long filenames on ISOs
Windows does not understand Rock Ridge (the equivalent extention in
Linux-land), and we need the full filenames to be available on Windows
(because of the assumption that the signature will have the full
eos...img.asc filename). This also means the installer binary appears
correctly as `endless-installer.exe`, not as `ENDLESS_INSTALLER.EXE`.

https://phabricator.endlessm.com/T14309