write-live-image: add Autorun.inf
On USB sticks, only the 'label' and 'icon' keys are used, which is
still better than nothing: "Endless-OS" is a better volume label than
"eoslive", and having the logo is nice.

On optical media, the 'open' key gives a "Run endless-installer.exe"
prompt when the disc is inserted.

The [Content] section disables scanning for media. Why not?

https://phabricator.endlessm.com/T14986