eos-diagnostics: Split coredump support off behind a --coredumps arg
It doesn’t really belong with hardware reporting, but does need to be
run as root, so split it off into a separate command line argument. This
requires a little reworking of the argument handling, and the
introduction of another tag to the diagnostics array.

Coredumps can now be dumped by running:
   sudo eos-diagnostics --coredumps
or along with the hardware info (but in a separate tarball):
   sudo eos-diagnostics --coredumps --hardware

Note that the systemd journal will be included in both tarballs, as it
contains relevant data for both use cases.

Signed-off-by: Philip Withnall <withnall@endlessm.com>

https://phabricator.endlessm.com/T15177