eos-write-live-image: optimisation for uncompressed .img source file
If we are making an ISO, we uncompress the image file into a tmpdir
before creating the squashfs. In the case the source image is
uncompressed, eos-write-image ends up being a used as a slow version
of cp. This optimisation attempts to hardlink the file in the hope
that the source image and tmpdir are on the same device, but otherwise
falls back to the previous code.