ISO: Do not use fragments for the squashfs image
GRUB has a (still unknown) bug that causes the ISO to not boot when the
number of fragments is >0. While investigating on the grub side, we fix
this avoiding using fragments at all for the squash image.

Signed-off-by: Carlo Caione <carlo@endlessm.com>
(cherry picked from commit da71141c101faab57345e19b0c3fc04286a162fe)
Signed-off-by: Will Thompson <wjt@endlessm.com>