write-live-image: add --delete-os-image flag
Once the SquashFS image is created, we don't need the original image at
all. This is intended for use on the image builder, where there is not
enough disk space for the uncompressed image, the SquashFS image, and
the ISO image all at the same time.

(Note that we already delete the uncompressed copy of the image in
$DIR_IMAGES_ENDLESS. In the image builder case, this is a hardlink to
$OS_IMAGE.)

https://phabricator.endlessm.com/T15668