ISO: Change partition type 0x83->0x00 for broken UEFI
Some broken UEFI systems refuse to recognize the ISO as bootable when
the main MBR partition has type 0x83 instead of type 0x00. Since we
cannot change this from xorriso we change the partition type directly
in the ISO image.

We do this by using `dd` instead of `sfdisk` since the latter complains
with `Start sector 0 out of range.` when trying to make the change.

Signed-off-by: Carlo Caione <carlo@endlessm.com>