Revert "ISO: Make El Torito img path explicit"
Now that we have verified that the only needed change to have the ISO
recognised by all the UEFI firmwares is the change in the type of the
main partition, we can avoid shipping the ESP partition in the ISO
itself.

With this patch we link the El Torito image to the ESP partition
instead of using the ESP image file on the ISO partition.

This reverts commit 2ddb11415ec56349272552e02dacd411f1552709.

Signed-off-by: Carlo Caione <carlo@endlessm.com>