eos-download-image: Use python requests instead of wget
Use the python requests library for downloading rather than using wget.
This has a few advantages:

 * By using a single requests session, the CloudFront connection
   automatically stays alive, speeding up subsequent requests.

 * Both the --timestamping and --continue behaviors can be supported
   without becoming incompatible with other wanted options.

 * The request to the server can be better controlled. This will be used
   later to request byte ranges.

 * A separate process isn't forked.

The downloading is all managed from the download() function. A simple
percentage progress is provided when the stderr target is a terminal
rather than a log file.

This adds new dependencies on python3-requests and python3-dateutil.
However, python3-requests is already implicitly included in the core and
python3-dateutil is only 173 kB.

https://phabricator.endlessm.com/T16099