eos-download-image: Use requests to download uncompressed index
Requests automatically adds Accept-Encoding: gzip to the request and
manages the decompression if compressed bytes are sent back. This means
that the the uncompressed index can be downloaded with no additional
bytes sent over the wire. Furthermore, the builtin requests json()
method can be used for decoding rather than feeding the bytes through
StringIO to json.load().

https://phabricator.endlessm.com/T16099