eos-download-image: Break up request when exceeding maximum size
CloudFront has an arbitrary 20 GB maximum limit for downloads. See
http://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/cloudfront-limits.html.
However, files can still be downloaded using Range requests below this
limit. Break up the request into multiple parts if the file size exceeds
the 20 GB limit.

Unfortunately, CloudFront also returns a 403 even when doing a HEAD
request on these files, so the size and modification time can't be
determined. In that case, they're supplied from the index as a hint.
Normally the downloader will prefer the information from the HEAD
request when that's available.

https://phabricator.endlessm.com/T16099