eos-write-{image,installer}: only partprobe target device
If the machine has an Endless OS ISO inserted into an optical drive,
global partprobe is unhappy; but probing the specific device we just
wrote to is fine (even if we wrote an ISO with eos-write-image: a
different partition table is used on non-optical media).

This is the same error squashed in 706efa2 for eos-write-live-image.

https://phabricator.endlessm.com/T15598