Install eos-b43fw-install
This is a script which downloads Broadcom proprietary Linux drivers from
the web, extracts the firmware files needed by the open source drivers
from the proprietary drivers, and install the firmware files in the
right place.

eos-b43fw-clean is also provided by the same package to uninstall these
files, if needed.

https://phabricator.endlessm.com/T12154