eos-list-users: continue if mounting a partition fails
If a partition can't be mounted, it might be due to a bug, or it might
be something wrong with that partition. Maybe the user has a hard drive
with a partition of type "basic data" but without a filesystem on it? In
particular, this happens when running this script from a combined USB,
because the physical eoslive partition (eg /dev/sdb1) cannot be mounted
directly. It's better to continue and list users on other partitions
even if we show an error.

https://phabricator.endlessm.com/T17560