eos-list-users: skip read-only partitions
Without this change, when booted from a combined USB, output like the
following is shown:

    FUSE exfat 1.2.3
    fuse: mount failed: Device or resource busy
    Can't list users on /dev/sdb1
    Users on /dev/mapper/endless-image3:

The first three lines are due to trying to mount /dev/sdb1, the image
host partition, which cannot be mounted because the image file within it
is mapped to /dev/mapper/endless-image. The final line is due to
enumerating users on the booted live system -- but because the image is
pristine, it will never contain users.

The primary purpose of this script is to determine what parameters to
pass to eos-reset-password. A partition must be writeable if we are to
reset a password within it, so I don't see an obvious downside to
skipping read-only partitions.

https://phabricator.endlessm.com/T17560