eos-stage-ostree: add support for demo and prod stages
Like dev, we set demo to check for an update every time
the updater runs.  Prod, on the other hand, follows the
normal production check for updates every two weeks.

On dev, we stick to a single product.major version,
to avoid ping-ponging when two major versions are
in development at the same time.  For demo and prod,
however, we upgrade to whatever is the latest OSTree
within the product series (e.g., eos3).

https://phabricator.endlessm.com/T17676