eos-stage-ostree: start the autoupdater at the end
Restart the autoupdater timer, and try to update now.
Note that downgrades (e.g., from demo to previous prod release)
won't happen automatically.  They would have to be done via
`ostree admin upgrade --allow-downgrade`, but that should be
a manual step if someone really wants to downgrade.

Note that this is a change in the behavior of eos-convert-system-qa,
as it previously left the autoupdater timer disabled.

https://phabricator.endlessm.com/T17676