eos-write-image: detect removable media automatically
The heuristics available from lsblk to detect removable media are:

* RM: matches the 'removable' attribute in sysfs. This is 1 for all
  USB sticks I have handy, 0 for the removable SD card inserted
  into my laptop, and 0 for a soldered-in MMC in at least one other
  laptop.
* HOTPLUG: 1 if RM is 1, or if the device chain includes a USB,
  IEEE1394, PCMCIA, MMC or CCW device. This is 1 for all devices
  mentioned above.

As discussed at https://phabricator.endlessm.com/T17871#409848
https://bugs.freedesktop.org/show_bug.cgi?id=49844 and elsewhere, these
heuristics are imperfect but we cannot do much better.

If we falsely believe a fixed device (eg soldered-on MMC) to be
removable, we will rely on the firmware loading the default boot path
\efi\boot\bootx64.efi for every boot. (This is what we did for all
systems prior to Endless OS 3.1.)

If we falsely believe a removable device (eg USB stick, or removable
MMC!) to be fixed, booting it on an EFI system erases the system's boot
configuration.  This has happened in the wild and is particularly bad on
dual-boot systems, since Windows' \efi\boot\bootx64.efi knows how to
boot Windows, but not (of course) Endless OS.

I think the latter is more dangerous, so let's err on the side of
caution and use HOTPLUG.

https://phabricator.endlessm.com/T17871