eos-rollback: Rollback even if the old deployment is booted
Rather than switch to the deployment that is not currently booted,
we change the default boot order.  This way, if the user has to
go into the GRUB menu to boot into the previous deployment,
he/she can still use eos-rollback to persist this old deployment.

Each call to this script rotates the boot order for the available
deployments, putting the first deployment to the end of the list.
In the case of Endless OS, there are only two deployments,
so we effectively toggle between the two available versions.

The implementation is migrated from awful command-line parsing
in bash to a clean python script provided by Dan Nicholson.

With the python API, we no longer have to redeploy, so the process
is much faster.

https://phabricator.endlessm.com/T18539