eos-write-live-image: remove --iso flag
This now lives in the OS image builder. It was never possible to run
this on an unconverted Endless OS system, since several of the
dependencies are not included in the ostree.

The eos-tech-support-iso metapackage existed to pull in the dependencies
for the --iso flag, so it too is now unneeded.

https://phabricator.endlessm.com/T17203