eos-write-live-image: consistently indent --help
This aligns the Options: section with the Usage: and Arguments:
sections, which is more pleasing to my eye. Adding a space before the
long options, and aligning --debug with the other long options, also
looks more pleasing and matches the --help output of a few other tools I
looked at.