eos-fix-ostree-repo: Workaround broken list_objects in old ostree
Prior to ostree-2017.1, the GI annotation for ostree_repo_list_objects
was wrong, making it unusable from bindings[1]. This is tricky to detect
since the version checking was not added until ostree-2017.3. However,
in the same commit, ostree_repo_list_commit_objects_starting_with was
fixed so that the out_commits parameter was marked properly.

Use the direction of this argument as a proxy to know when
ostree_repo_list_objects will work. Otherwise, make our own list_objects
implementation and monkey-patch it into the Repo class.

1. https://github.com/ostreedev/ostree/commit/300752e5

https://phabricator.endlessm.com/T19420